Move seqlens/offsets generation to mask descriptor
